requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
sentence-transformers
pgvector
redis
python-dotenv
python-jose[cryptography]
openai
passlib==1.7.4
bcrypt==4.0.1
boto3
python-dotenv
tiktoken